requirements.txt
# python==3.12.3 specified in Dockerfile
flask==3.0.3
snowflake-connector-python==3.12.0